Move metrics resolvers into dedicated modules
