+++ back/boxtribute_server/graph_ql/resolvers.py
@@ -81,13 +81,6 @@ from ..models.definitions.tag import Tag
 from ..models.definitions.tags_relation import TagsRelation
 from ..models.definitions.transfer_agreement import TransferAgreement
 from ..models.definitions.unboxed_items_collection import UnboxedItemsCollection
-from ..models.metrics import (
-    compute_moved_stock_overview,
-    compute_number_of_beneficiaries_served,
-    compute_number_of_families_served,
-    compute_number_of_sales,
-    compute_stock_overview,
-)
 from .bindables import (
     base,
     beneficiary,
@@ -96,7 +89,6 @@ from .bindables import (
     distribution_event,
     distribution_events_tracking_group,
     distribution_spot,
-    metrics,
     organisation,
     packing_list_entry,
     product,
@@ -398,19 +390,6 @@ def resolve_shipments(*_):
         authorized_bases_filter(Shipment, base_fk_field_name="source_base"),
         authorized_bases_filter(Shipment, base_fk_field_name="target_base"),
     )
-
-
-@query.field("metrics")
-@convert_kwargs_to_snake_case
-def resolve_metrics(*_, organisation_id=None):
-    # Default to current user's organisation ID
-    organisation_id = organisation_id or g.user.organisation_id
-    # Non-god users are only permitted to fetch their organisation's metrics, the god
-    # user however can access any organisation's metrics
-    authorize(organisation_id=organisation_id)
-
-    # Pass organisation ID to child resolvers
-    return {"organisation_id": organisation_id}
 
 
 @tag.field("taggedResources")
@@ -1178,41 +1157,6 @@ def resolve_location_boxes(location_obj, _, pagination_input=None, filter_input=
         selection = Box.select().join(Product)
     return load_into_page(
         Box, filter_condition, selection=selection, pagination_input=pagination_input
-    )
-
-
-@metrics.field("numberOfFamiliesServed")
-def resolve_metrics_number_of_families_served(metrics_obj, _, after=None, before=None):
-    return compute_number_of_families_served(
-        organisation_id=metrics_obj["organisation_id"], after=after, before=before
-    )
-
-
-@metrics.field("numberOfBeneficiariesServed")
-def resolve_metrics_number_of_beneficiaries_served(
-    metrics_obj, _, after=None, before=None
-):
-    return compute_number_of_beneficiaries_served(
-        organisation_id=metrics_obj["organisation_id"], after=after, before=before
-    )
-
-
-@metrics.field("numberOfSales")
-def resolve_metrics_number_of_sales(metrics_obj, _, after=None, before=None):
-    return compute_number_of_sales(
-        organisation_id=metrics_obj["organisation_id"], after=after, before=before
-    )
-
-
-@metrics.field("stockOverview")
-def resolve_metrics_stock_overview(metrics_obj, _):
-    return compute_stock_overview(organisation_id=metrics_obj["organisation_id"])
-
-
-@metrics.field("movedStockOverview")
-def resolve_metrics_moved_stock_overview(metrics_obj, _, after=None, before=None):
-    return compute_moved_stock_overview(
-        organisation_id=metrics_obj["organisation_id"], after=after, before=before
     )
